feat : 수정 및 id 검증 로직 추가

+++ src/main/java/org/sopt/assignment1/DiaryService.java
@@ -16,10 +16,27 @@ public class DiaryService {
     }
 
     void deleteDiary(final String id) {
+        if(!isExisted(id))
+            throw new NullPointerException("존재하지 않는 id 입니다.");
         diaryRepository.delete(id);
     }
 
-    void validateId(final Long id) {
+    void patchDiary(final String id, final String body) {
+        if(!isExisted(id))
+            throw new NullPointerException("존재하지 않는 id 입니다.");
+        diaryRepository.update(id ,body);
+    }
 
+    boolean isExisted(final String id) {
+        List<Diary> diaryList = getDiaryList();
+        if (diaryList.isEmpty()) {
+            return false;
+        } else {
+            for (Diary diary : diaryList) {
+                if (diary.getId().equals(Long.parseLong(id)))
+                    return true;
+            }
+        }
+        return false;
     }
 }
